# specs/001-hackathon-ai-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: What If Visualizer API
  version: 0.1.0
  description: API contracts for prompt-driven image/video generation, refinement, publishing, and feed.
paths:
...
  /api/generations:
    post:
      summary: Create a new generation job (image or video)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequest'
      responses:
        "202":
          description: Job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generation'
...
  /api/publish:
    post:
      summary: Publish a generation to the community feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        "201":
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPost'
...
components:
  schemas:
    CreateGenerationRequest:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum: [image, video]
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
        guidance:
          type: string
          nullable: true
        language:
          type: string
          enum: [en, zh-CN, ja, unknown]
          default: unknown
        refinementOf:
          type: string
          nullable: true
      required: [type, prompt]
...
    PublishRequest:
      type: object
      additionalProperties: false
      properties:
        generationId:
          type: string
        visibility:
          type: string
          enum: [public, unlisted]
        altText:
          type: string
          nullable: true
        captions:
          type: string
          nullable: true
      required: [generationId, visibility]
    Generation:
      type: object
      properties:
        id:
          type: string
        promptId:
          type: string
          nullable: true
        type:
          type: string
          enum: [image, video]
        status:
          type: string
          enum: [queued, running, complete, failed]
        model:
          type: string
        refinementOf:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        error:
          type: string
          nullable: true
        asset:
          $ref: '#/components/schemas/MediaAsset'
          nullable: true
    MediaAsset:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        storagePath:
          type: string
        format:
          type: string
          description: File format, e.g., png or mp4
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        durationSec:
          type: number
          nullable: true
        altText:
          type: string
          nullable: true
        captions:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [private, unlisted, public]
    CommunityPost:
      type: object
      properties:
        id:
          type: string
        generationId:
          type: string
        authorId:
          type: string
        promptSummary:
          type: string
        thumbnailUrl:
          type: string
          format: uri
        publishedAt:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [public]
...